product

+++ frontEnd/src/components/routes/Auth.tsx
@@ -1,14 +1,18 @@
 import { queries } from '../../features/user/api/queries'
 import Storage from '../../utils/storage'
 import { Navigate, Outlet } from 'react-router-dom'
-
-const Auth = () => {
-  const { isError } = queries.useMe()
+import SomethingWentWrong from '../feedback/SomethingWentWrong'
+type role = {
+  AllowedRole: string[]
+}
+const Auth = ({ AllowedRole }: role) => {
+  const { data, isError, isLoading } = queries.useMe()
   const token = Storage.getToken()
-
+  if (isLoading) return ;
   if (token) {
     if (!isError) {
-      return <Outlet />
+      if (AllowedRole.includes(data?.role)) return <Outlet />
+      else return <SomethingWentWrong text="403 Forbidden" />
     } else {
       Storage.removeToken()
       return <Navigate to="/login" />
